Make StrategyHandler improvements
Creating StrategyUtils starting with some methods to read
from the position data files written daily. These methods
allow StrategyHandler to dynamically read in the parameters
found in strategy_parameters.json and run checks on the
suggested trade found in the signal.

Minor changes to OrderTypes, ParserHandler, and MarketDataConstants.

+++ src/main/java/com/paperstreet/strategy/OrderTypes.java
@@ -14,7 +14,7 @@ public class OrderTypes {
      * @param quantity amount to trade
      * @return a market order
      */
-    public static Order MarketOrder(String action, double quantity) {
+    public static Order MarketOrder(String action, int quantity) {
         Order order = new Order();
         order.action(action);
         order.orderType("MKT");
